+++ riscv/src/register/mcause.rs
@@ -1,6 +1,7 @@
 //! mcause register
 
 pub use crate::interrupt::Trap;
+use crate::{CoreInterruptNumber, ExceptionNumber};
 
 read_only_csr! {
     /// `mcause` register
@@ -37,6 +38,24 @@ read_only_csr_field! {
 }
 
 impl Mcause {
+    /// Creates an `Mcause` value representing the given core interrupt source.
+    ///
+    /// The [interrupt bit](Self::IS_INTERRUPT_MASK) is set and the `code` field
+    /// is set to the interrupt number.
+    #[inline]
+    pub fn from_interrupt<I: CoreInterruptNumber>(interrupt: I) -> Self {
+        Self::from_bits(Self::IS_INTERRUPT_MASK | interrupt.number())
+    }
+
+    /// Creates an `Mcause` value representing the given exception source.
+    ///
+    /// The interrupt bit is clear and the `code` field is set to the exception
+    /// number.
+    #[inline]
+    pub fn from_exception<E: ExceptionNumber>(exception: E) -> Self {
+        Self::from_bits(exception.number())
+    }
+
     /// Returns the trap cause represented by this register.
     ///
     /// # Note
@@ -56,5 +75,29 @@ impl Mcause {
     #[inline]
     pub fn is_exception(&self) -> bool {
         !self.is_interrupt()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::interrupt::machine::{Exception, Interrupt};
+
+    #[test]
+    fn test_mcause() {
+        let m = Mcause::from_interrupt(Interrupt::MachineExternal);
+        assert!(m.is_interrupt());
+        assert!(!m.is_exception());
+        assert_eq!(m.code(), Interrupt::MachineExternal as usize);
+        assert_eq!(
+            m.cause(),
+            Trap::Interrupt(Interrupt::MachineExternal as usize)
+        );
+
+        let m = Mcause::from_exception(Exception::Breakpoint);
+        assert!(m.is_exception());
+        assert!(!m.is_interrupt());
+        assert_eq!(m.code(), Exception::Breakpoint as usize);
+        assert_eq!(m.cause(), Trap::Exception(Exception::Breakpoint as usize));
     }
 }